pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aiworkos-backend"
version = "0.1.0"
description = "Production-ready FastAPI backend for an AI Work Operating System built on the OpenAI Agents SDK."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "aiosqlite>=0.20.0,<1.0.0",
  "fastapi>=0.115.0,<1.0.0",
  "mcp>=1.6.0,<2.0.0",
  "openai-agents>=0.0.14,<1.0.0",
  "pydantic-settings>=2.4.0,<3.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
]

[tool.setuptools.packages.find]
include = ["backend*"]

[tool.setuptools.package-data]
backend = ["py.typed"]
